all: populate record tag after JSON unmarshal
Due to the way Go initializes structs their fields cannot have a
default value. This means we use exported constructors to populate
them with hidden fields. When we unmarshal (via FileFromJSON) those
hidden files are not populated, so we have to sneak their values in
whenever someone uses json.Unmarshaler

Fixes: https://github.com/moov-io/wire/issues/104